exp/requirements.txt
absl-py==1.3.0
aiofiles==22.1.0
anyio==3.6.2
appdirs==1.4.4
argcomplete==2.0.0
astunparse==1.6.3
audioread==3.0.0
boto3==1.26.8
botocore==1.29.8
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cityscapesScripts==2.2.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.6
crc32c==2.3
cryptography==38.0.3
cycler==0.11.0
dacite==1.6.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
eventlet==0.33.1
fiftyone==0.18.0
fiftyone-brain==0.9.2
fiftyone-db==0.4.0
flatbuffers==1.12
fonttools==4.38.0
future==0.18.2
gast==0.4.0
gcloud==0.18.3
glob2==0.7
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
graphql-core==3.2.3
greenlet==2.0.1
grpcio==1.34.1
h11==0.12.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.21.0
httpx==0.23.0
humanfriendly==10.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
imageio==2.22.4
importlib-metadata==5.0.0
intel_tensorflow_avx512==2.9.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kaleido==0.2.1
keras==2.9.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
mongoengine==0.24.2
motor==3.1.1
ndjson==0.3.1
networkx==2.8.8
nibabel==3.2.1
ninja==1.11.1
numba==0.56.4
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
patool==1.12
Pillow==9.3.0
pkg_resources==0.0.0
plotly==5.11.0
pooch==1.6.0
pprintpp==0.4.0
priority==2.0.0
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyelftools==0.29
pymongo==4.3.2
pyparsing==3.0.9
pyquaternion==0.9.9
python-dateutil==2.8.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
retrying==1.3.3
rfc3986==1.5.0
rsa==4.9
s3transfer==0.6.0
samplerate==0.1.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.11.0
sox==1.4.1
soxbindings==1.2.3
sse-starlette==0.10.3
sseclient-py==1.7.2
starlette==0.20.4
strawberry-graphql==0.138.1
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==1.1.0
tfrecord==1.14.1
threadpoolctl==3.1.0
tifffile==2022.10.10
toml==0.10.2
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
universal-analytics-python3==1.1.1
urllib3==1.26.12
voxel51-eta==0.8.1
Werkzeug==2.2.2
wrapt==1.12.1
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.10.0
